requirements.txt
Flask==2.3.2
# The main web framework for building your application.

pandas==2.2.2
# For data analysis and manipulation (useful for student performance analytics).

numpy==1.26.4
# Fundamental package for numerical computations (used by pandas and scikit-learn).

matplotlib==3.8.3
# For creating static, animated, and interactive visualizations.

seaborn==0.13.2
# Statistical data visualization built on top of matplotlib.

Flask-SQLAlchemy==3.1.1
# Adds SQLAlchemy support to Flask for ORM-based database access.

scikit-learn==1.4.2
# Machine learning library for predictive analytics and data mining.

Jinja2==3.1.3
# Templating engine used by Flask for rendering HTML templates.

gunicorn==21.2.0
# Production WSGI server for deploying Flask apps (used on platforms like Render/Heroku).

python-dotenv==1.0.1
# Loads environment variables from a .env file (optional, but useful for config management).

Flask
# (Duplicate, but ensures Flask is installed if version not specified above.)

Flask-Mail
# For sending emails (used for OTP and password reset).

Flask-SQLAlchemy
# (Duplicate, but ensures SQLAlchemy support for Flask.)

Werkzeug
# WSGI utility library used by Flask.

itsdangerous
# Used by Flask for securely signing data (e.g., session cookies).

click
# Command-line interface creation toolkit used by Flask.

Jinja2
# (Duplicate, but ensures template engine is present.)